[refactor]: enum 파일 구조 변경 (#65)
