+++ app/src/main/java/com/kuit/ourmenu/data/model/base/type/MenuFolderIconType.kt
@@ -1,4 +1,4 @@
-package com.kuit.ourmenu.data.model.menuFolder.response
+package com.kuit.ourmenu.data.model.base.type
 
 enum class MenuFolderIconType {
     ANGRY,
