+++ app/src/main/java/com/kuit/ourmenu/ui/menuFolder/viewmodel/MenuFolderDetailViewModel.kt
@@ -3,7 +3,7 @@ package com.kuit.ourmenu.ui.menuFolder.viewmodel
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.kuit.ourmenu.data.model.menuFolder.response.MenuFolderDetailResponse
-import com.kuit.ourmenu.data.model.menuFolder.response.SortOrderType
+import com.kuit.ourmenu.data.model.base.type.SortOrderType
 import com.kuit.ourmenu.data.repository.MenuFolderRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
